refactor(events): move attendance type in a separate file
