+++ frontend/src/common/enums/attendance-type.enum.ts
@@ -1,0 +1,4 @@
+export enum AttendanceType {
+  SIMPLE = 'simple',
+  MENTION = 'mention',
+}
